Add very basic tests of index() function
Although this causes the test suite to slow down a lot, it at least
tests the code and hence gives some support for later refactoring
efforts.  The indexing operation takes roughly a minute on my box, so
this should not be a problem for other dev machines.  The comment about
the indexing operation taking about half an hour seems as if it could be
outdated.