Add tests of the modlist() function
This function requires the index to have been built, hence we build the
index again in a new temporary directory to ensure a clean slate before
starting the tests.  This could be improved upon by not indexing an
entire Perl installation as part of the tests, however the code needs to
be refactored to allow that and these tests are a step to enabling
such future refactorings.